New scripts to aid in testing; on one machine you should be able to run
	sh tests/llmount-server.sh

and on another machine:

	sh tests/llmount-client.sh
	mount -t lustre_light -o device=3 none /mnt/obd

and do things like:

	ls -ld /mnt/obd
	chmod 777 /mnt/obd

This works now--even over the socknal, on an SMP kernel (although I haven't
tested it on an SMP machine)

You'll need to edit the SERVER variable in llmount-client.sh.  If you want
to get really fancy, you'll make it a command-line option.