Add journal wrapping calls to all of the mds_reint operations.  We also use the mds_fs_setattr method to change the fields in the inode, as this will properly mark the inode dirty and ensure it is written to disk.
This fixes the problem of updates to the MDS inode not being written to
disk because the dentry and inode are being discarded too quickly.

The journal wrapping of all reint operations will make the addition of
calls to atomically update last recieved trivial, so it is time well spent.