Allow obdctl to use "$OBDDEV" to resolve a device number in setup.
This required changing the NAME2DEV ioctl so that it didn't change the
currently selected device when it was resolving a name.  Now the NAME2DEV
ioctl only resolves the name, and obdctl selects the returned device
explicitly (to user-space the "name2dev" command works exactly the same).

Cleaned up setup scripts to remove last vestiges of hard-coded device numbers.
Instead we use "setup $OBDDEV" (note that the '$' must be escaped from the
shell if using it in a shell script).