Send last_rcvd values around when talking to the MDS.  The MDC gets the last_{rcvd,committed,xid} values on mdc_connect, but doesn't yet do anything with this new data except print it to the debug logs.
A select number of MDS operations get last_{rcvd,committed} values sent
in the reply (mds_body) - create, getattr, open.  It is not totally
clear to me how to add in the mds_body to an RPC reply if it doesn't
already exist, so there is a little more work to do there.

At connect and reint time, client "UUIDs" are looked up and handled
appropriately for new and existing clients.  Currently, since the
RPCs don't actually contain any UUID values, all updates go to UUID "",
which is enough for testing, and should "just work" when UUIDs appear.