Fix problem with multiple connections to the same local device.
The problem was that we stored the remote export addr+cookie in the
device array.  When we disconnected any connection, it would drop the
remote export from the most recent connection.

What we do now is store the remote export addr+cookie in the local
export/import struct and always use that (the old mdc_connh and osc_connh
are gone).  We could do some further cleanup of the API at this point,
but I will leave it until there is some discussion on the matter.  I was
able to successfully do multiple simultaneous connect and disconnect steps,
as well as run multi-thread test_getattrs with the change.  I'm just doing
a full runregression-net test.

I'm still not 100% confident that this is what Peter wants, so I tagged
the pre-commit tree with "before_rconn" to make it easy to generate
a diff (I will also tag the post-commit with "after_rconn").