Add some basic data integrity checking to obdecho.
This puts the offset and objid into the first 16 bytes and last 16 bytes
of the bulk transfer.  These are in HTON__u64() format.

The lctl command for test_brw now takes an objid instead of an obdo count,
so that you can (potentially) use lctl with test_brw on a real OBD instead
of just obdecho.  That has not been tested yet.