Add module loading support to lconf. By default, lconf will load and unload the modules needed based on what devices are configured for a node.
The path to load modules from is determined based on the directory lconf is
run from. If a Makefile is found, the it is assumed lconf is in lustre/utils
and modules will be searched for in ../../lustre and ../../portals.

Module support can be turned off with --nomod, if desired.

Use option --gdb  to create a gdb module script. Lconf will print the path
of the script and pause for few seconds.