We already set last_committed and last_xid for all packets, no need to do this for getstatus explicitly.
Store an xid in last_xid and not last_rcvd for reply packets.  Mike please
verify this.  I think it will help resolve strangeness I have observed in
the cleanup of expired requests on the clients, but that still needs some
investigation.

Add a spinlock to protect 64-bit increment of mds_last_rcvd, and also
make sure we only use each transaction value a single time.  This
would probably be a very rare issue for recovery, but may as well get
it now.