Mike's fix for the missing exp_connection oops caused us to not remove old exports from the export list.
It also seems wrong, somehow, to be using a lock on the ptlrpc connection
to manage items that are linked to the obd_device struct.  It would
probably make more sense to protect the export list with a lock on the
obd_device.  This is doubly true, because class_new_export() is adding
the export to the obd_device without holding any lock, because there is
no ptlrpc_connection there yet.