Change internal statfs calls to pass struct obd_statfs as a parameter instead of the Linux struct statfs.  This allows us to have 64-bit clean statfs data all the way to the VFS, and avoids problems like block count overflows in intermediate math in LOV and such.
We still need to resolve the VFS interface to statfs, as it limits the
output to 16TB on 32-bit platforms (4G x 4kb blocks).  The likely
answer is to shift the blocksize up, and the blocks counts down until
we fit in under 4G total blocks.  For MCR this will be around a blocksize
of 64kB which is nice to specify as the block size anyways.