b=605627 Fixes two bugs related to 605627 that have been present since we started correctly reusing inodes.
- We were giving up locks at file close time, but not throwing away file data
pages.  Fixed, by adding a flag to the blocking callback.  It's now called
under two different circumstances: when a lock needs to be given up
(LDLM_CB_BLOCKING) and when a lock is about to be freed (LDLM_CB_DYING)

- We were not refreshing inode attributes (notably size) correctly.  I brute
force this by always calling ll_file_size() in ll_inode_revalidate, but this
needs some obvious immediate refinement.

As an aside, I noticed that the DLM API documentation gives almost no mention
to the arguments or calling conventions of the callback functions.