The old patch was failing because sometimes the inodes were gone by the time the callback was called with CB_DYING.  I also realized that the client would believe that it had valid data for too long.
Instead of an extra callback when the lock is freed, I moved that extra
callback to just before we cancel the lock.  Because there's two paths for that
now (cancelling with server notification and cancelling without), I had to
introduce _another_ l_flag, which indicates whether we've done this call.  Oh
well.  It works much better now.