Fix (or a workaround that will hide the real bug) for bug 322.
We are calling mdc_blocking_ast() to clean up an inode that is already
being cleaned up elsewhere.  This is the result of ldlm_cli_cancel_unused()
being called from two places at the same time:
- ll_prepare_recovery()->prepare_mdc()->ldlm_cli_cancel_unused()
- ll_clear_inode()->mdc_cancel_unused()->ldlm_cli_cancel_unused()

There is still a small race where the ll_clear_inode() RPC could have
completed and "inode" is garbage, but since we are in recovery this
is highly unlikely.