The code at the end of mds_handle is extremely dangerous unless we go ahead with our plan to single-thread the MDC, so I've single-threaded it.
This was most easily done by moving the XID allocation into ptlrpc_queue_wait,
which had only minimal fallout: I had to allocate a new XID for MDS_READPAGE,
instead of getting it from the request after prep_req.  To avoid leaving holes
in the MDS XID space, I added a second xid (imp_last_bulk_xid) to the import,
which is safe because these packets go to different portals anyways.  I
communicate this to the MDS in the 'blocks' field of the mds_body, the only
64-bit field that was left in the mds_readpage usage of mds_body.  Then it was
a simple matter of adding down/up around all of the MDC's ptlrpc_queue_waits.

Robert: you will probably need to use this variable on the OSC, too, unless
you don't mind holes in the XID space.