Most of the MDS unlink handling.
For some reason using the passed mds_last_rcvd_cb as a parameter to
mds_finish_transno() as an arg to fsfilt_set_last_rcvd() doesn't work
very well (crashes UML), yet passing the same function pointer directly
to fsfilt_set_last_rcvd() seems to work OK.

AAAAAAARRRRRRGGGGGGGHHHHHHH...  I see it now - prototypes for functions
that are declared directly in another file, which are now WRONG!!!!!!!!
I hate it when that happens.  To be fixed in next checkin.