Merge from b_devel -> b_orphan   DEVEL_ORPHAN_UPDATE_PARENT_20030528 -> DEVEL_ORPHAN_UPDATE_PARENT_20030604
This is a non-functional commit, with osc/osc_request.c not compiling because
it is missing the "llog_commit_master" data struct, and does not have any
commit threads started yet.  I'm beginning to think the right thing to do
here is to create a single global llog_commit_master struct for the OSC, and
start a single thread at module setup time, and clean it all up at module
unload time, but I could be convinced otherwise.