b=1497 Completely by accident, I may have found the source of this bug.  My theory is as follows:
When a client metadata lock is enqueued, it is first requested on the parent
directory (for the future, when we will grant that writeback lock).  Today, the
client is either refused the lock entirely, or is granted a lock on the child.
This is the case in unlink.

In the old code, unlink got a lock on this new child resource, but we never did
an iget() for the inode on the client.  So the lock continued to point to the
parent inode, even though it was enqueued on the child resource.

It could happen, however, that the parent inode is subsequently pushed out of
the inode cache (by a lock revocation or memory pressure).  Normally this would
first result in the cancellation of any existing locks for that inode, but
because this lock is actually on the _child_ resource, the lock would remain
enqueued but the inode pointer freed.  Later, when this lock gets freed, we
would dereference the now-bad inode pointer.  Et voila.

...

The kicker here is that there is absolutely no reason for unlink to get a lock
at all--in fact, the correctness of the protocol only requires that the MDS get
this lock, and not that it be returned to the client.  This patch makes unlink
just like most other operations, which do not acquire locks.