Landing statfs cache code. - preliminary for prealloc (bug 974) - preliminary for QOS (bug 855) - speeds up parallel df - speeds up lprocfs
Lprocfs underwent a bunch of cleanup (mostly from Intel, partly from me)
to remove lots of macros and strangeness w.r.t. per-target statfs values.

Still open for discussion is how old the cached statfs values can be.
I arbitrarily picked 1 second for client statfs, lprocfs, and target
statfs caches.  We may want to change that, or allow it to be tunable.

Note that this code is different than that in bug 855, in that the
max_age parameter is an absolute age in jiffies, whereas the patch
in bug 855 used a relative time in seconds.  The thing I'm trying
to avoid is if we say "I want a result up to 1 second old", but we
get delayed during an RPC and we don't know what "1 second old" is
related to.