This fixes the bug that new S_* debug subsystems were not being printed in debug logs.  The problem was a desynchronization between the kernel and userspace tools.
The old code would silently drop unknown subsystem messages.  Now a
warning is printed and they are included.

The subsystems are no longer a crazy twist of left-shifted integers, but
a normal bitmask.  There may have been a good reason for the craziness,
to try to fit a subsystem and type into one integer, but it's long gone.