b=1776 r=shaver Change the read_record and write_record APIs to return 0 on success and a negative error.  Any short read/write is reported as -EIO.
It used to return the number of bytes completed, which meant that the
error checking code had to be duplicated everywhere.  A short I/O in
either of these functions is always a failure.

(As a nice side effect, one of the orphan functions got this wrong, so
this gets us somewhat closer to having that infrastructure)