Remove "ERR" trap, which doesn't work in my UML, nor at PNNL. r=robert
Use "MOUNT" and "DIR" to be more in line with sanity.sh.  This also
allows us to run these tests in a subdir instead of the root, in case
it makes a difference (it doesn't currently).

Use checkstat instead of "ls" so that we can do some verification of results.

Don't pause for GDB debugging symbols to be loaded.