b=1803 r=shaver
New import state machine, as documented on the lustre wiki in
ImportStates.

A new function, ptlrpc_connect_import, performs all import connects
and moves the import from the DISCON state to either FULL, EVICTED,
REPLAY, or RECOVER, depending on the situation. Unlike the levels, the
states are now exact, and the request->rq_send_state much match the
import state to be sent.

Passes recovery/01, replay-small, and replay-dual.