A slightly hacked up version of Zach's patch to change the OBD API to use exports intead of handles.
It runs and cleans up normally -- as normally as llpmd ever did -- but
probably still contains some bugs.  These bugs will likely happen in
unused (cobd, ptlbd) or rarely taken (startup/cleanup error cases),
and will be of the type that leak an export or do an extra export_put.

I also reduced the obd_refcount sleep at cleanup to 1 second from 30.
I have never seen the count ever abort before the full 30 seconds were
up, and because of the unique configuration situation on b_llpmd, this
path is always taken during the cleanup of the LOV (because the MDS is
still holding an export reference)