Update checkstack program, to output more useful "stack_usage function name".
Add a "make checkstack" target which shows the current "worst offenders" list
of stack users, or the difference in this list since the last time that target
was run.

There is also a "make checkstack-clean" tarket which resets the "saved" stack
values so that the stack usage diff doesn't show old items.