Fix for missing open flags for vfs_intent-2.4.20-rh kernel patch. This change was also needed for the vfs_intent-2.4.20-hp patch.  b=1877 Also add bug 1344 revalidate loop fixes to some of the vfs_intent patches.
The diff is fairly large, so that the files in the 2.4.20-rh patch are
in the same order as in all of the other vfs_intent patches to make
comparison and maintenance easier.  Some other minor changes are made to
keep the patches consistent.

There is still a difference between vfs_intent-2.4.20-vanilla and -rh
in follow_down, but I don't know which one is the "keeper".