The easiest and least controversial part of the patch to eliminate the gymnastics associated with setting l_completion_ast on the client.
We used to set l_completion_ast when the lock is created, so that a
completion AST which arrives before the RPC reply will still be
handled correctly.  Later, we would NULL it out so that lock_enqueue's
call to grant_lock didn't needlessly execute again.  This was
race-prone and stupid.

Now ldlm_grant_lock takes a flag, so we can suppress needless calls
without this mess.