Stop storing new LOV EAs inline with the inode, and when we stat existing inodes that have LOV EAs inside the inode we convert them to proper EAs (whether external EA blocks or large-inode fast EAs).  The conversion part will be removed prior to Lustre Lite 1.0.
Anyone running an existing b_devel tree should be able to run

	find /mnt/lustre -type f -size +0

to do one-time moving of EAs.  After this is complete it is possible to do
properly run e2fsck on the MDS filesystem, assuming you don't have large
inode support enabled (which is being fixed as part of bug 841).

b=2097
r=braam