b=2109 Flesh out the async OBD IO API to be more user friendly.  llite behaviour should be the same, but easier to follow.  Each OBD layer gets its own struct of state to associate with a page in flight.  A struct of callback members is introduced to manage the communication between the obd async api providers (lov/osc) and the callers (llite/lov).
This is landing now, instead of after 1.0, to give liblustre a more reasonable
interface into the RPC concurrency engine in the OSC.  This mechanism is most
evident in the part of ll_commit_write() that falls back to sync IO.  O_DIRECT
will also want to use this.

These changes pass acceptance-small on real hardware.