Fix grant leak because we were sending "dropped" grant during read BRWs but the target wasn't doing anything with them.  This fixes the largest source of leak that prevented fsx from completing on a small filesystem.
There is still some small grant leak that means the OST is continually
increasing the client's grant, but the client keeps running out on its
side before consuming all of the grant that the OST has given it.  I'm
trying to thing of some way to have a self-correcting system, but the
fact that multiple RPCs can arrive/complete out-of-order makes this
difficult.

I had hoped that I could use the fact that we wait for all RPCs to
complete before doing a sync RPC without grant would allow the OST to
"reset" it's grant for that client (because it meant there was only
that RPC being sent), but this doesn't seem to be the case.
b=974