Fix racing write-vs-truncate grant accounting bug (osc_complete_oap() was being called from osc_send_oap_rpc() while truncate was in progress and all of the osc_exit_cache() calls were erronously being marked "sent").
This will be running fsx overnight, but seems promising at least (no leaks
so far).

Also includes a change in lop_makes_rpc() to start sending a write RPC when
we have just over an optimal-sized RPC for this file, instead of when we
hit 2x optimal RPC size.  This gets an RPC in flight early so that we don't
get close to running out of grant before the first RPC finishes.
b=974