b=2712
If mount is unable to connect to the config device to get the profile,
then it fails right away and doesn't try to recover.  (It will still
attempt recovery on the initial connect for the actual devices in the
config, though.)  If we need to revert this behavior for some reason,
all that is needed is to change the 0 back to 1 in the call to
lustre_process_log in lustre_fill_super.

Fixed the ptlrpc_connect_interpret to also invalidate the import when
we are not recoverying an initial connect.

Changed the test I just added to jive with this new behavior, and also
added another test suggested by Jacob.