Fix up catalogs so that log records pack evenly into each chunk.  This is required since catalogs are round-robin and entries might be added out-of-order so they always have to align properly.
To ensure the few sites that are (or have) run 1.2 catalog code will not
break, the catalog record MAGIC is changed and the name of the CATLIST
(now CATALOG) file is also changed.

To ensure CATALOG entries are 64-bit aligned, they are padded to 32 bytes
as llog_catid instead of llog_logid entries.
b=2733