Drop the filter export grants under the same lock as the removal of the export from the device from the export list.  This still isn't possible for --force disconnects, so we don't do filter_grant_sanity_check() at that time.
Also, change the LASSERTs to CERRORs in case this happens again.  First of
all to see if these errors are just transient during export disconnection
which means they are just overly-aggressive assertions, or if there is a
permanent accounting error that indicates some other problem.  Secondly, it
isn't possible to stop the use of existing export references, so there will
always be some small race during client eviction as completing operations
might update the export before it is destroyed.
b=2663