b=3885
THe hack on b1_2 branch to call set_info("mds_conn") on a specific OSC
instead of all of them was removed on HEAD, and this broke
mds_lov_synchronize.  Instead of introducing a new hack to do this,
for now this changes m_l_s to send the set_info() to all the connected
OSTs. This should be harmless.