Branch: b1_4 Move the MDC EA/cookie max size initialization later in the mount process, until we have established our permanent connections.  This avoids needless a bunch of client->OST connect+disconnect RPCs, and also fixes a bug where the client didn't get the reply for the first disconnect, and then hit a sanity check (target handle mismatch) on the server.
This will tie in directly to the bug 5707 fix, as that will additionally
store a "default" EA size to avoid allocating and sending a very large
(and otherwise useless) request buffer, just to have a place to store
the LOV EA for a MDS open replay.    Includes regression test.
b=5317
r=nathan (with minor fix to avoid the storage of lovdesc at all)