b=5981 r=adilger
SetPageDirty in 2.6 is a hangman's noose made of time bombs, with which to
shoot yourself in the foot.  It doesn't adjust page lists or counters, which
we now believe was almost certainly responsible for the two weeks of
mysterious "negative nr_dirty" hangs at NERSC and NOAA.

As an extra precaution, I also move the clear_page_dirty_for_io compat function
into the 2.4-only part.  If configure happened to ever enable it in 2.6
somehow, it would be a similar disaster (assuming that the compile didn't
fail).