b=4952 r=adilger Protocol change!  This commit breaks wire protocol compatibility.
connect messages now exchange another buffer, which currently contains only
flags.  This is used to implement read-only mounting, but will be used to fix
other issues as well.

The flags are exchanged with the MDT and OST, but currently only the MDT pays
any attention to them.  The places that check the disk are carefuly to check
the export for its read-only status.  Having this flag in the export gives us
the opportunity to also set it when the client doesn't request it, to force
read-only mounts as a security policy.

This commit also adds remount_fs method, which can make use of mdc_set_info to
change the rw/ro status at remount-time.  And osc_set_info should follow.