b=6377 r=adilger
Use obd_recovering instead of obd_recoverable_clients to check if the
obd is in recovery.

abort_recovery didn't set obd_recoverable_clients to 0 (which it now
does), so if recovery was aborted, the server would never ping-evict
clients (which can lead to extra exports, possibly requiring recovery
to be aborted).