- added more debug in osc. Currently two ideas are working for checks why osc does not give desired band:
(1) cache size is choosen kind of cumbersome way and not enough, what makes writing thread waiting for cache in osc_enter_cache(). Cache size curremtly does not depend on how good wire is able to send RPCs to remote node what seems to be an artificial limit in some cases. If this is so, the fix would be to adjust cache size depending on much it is needed. This however does not explain why second subtest suffers from slowdown whereas first one does not.

(2) due to some reasons (limited cache or limited grant from OST) many async RPCs are falled back to synchronous IO what makes substential slowdown.

Added debug on real cluster should show what is going on there. Idea that issue on server has a little chance to be true, as server uses direct IO and there is almost not ability to be slowed down due to some cache issues, etc.