b=9879 r=adilger
Test that we do not receive any plain lock requests from ibits-aware
clients.
Test that client and server have intersecting lock bits, and if not - server
drops IBITS compat flag. (if the flag is set and intersecting list of bits is
zero on client - just error message is issued).
We now also check that we do not return IBITS lock with sync callback to
liblustre clients.