b=9794 r=adilger
This patch uses a fast, high-quality PRNG internal to Lustre instead of the
glibc PRNG to avoid perturbing the glibc PRNG sequence for liblustre clients.
It also uses this PRNG for some uses in the lustre core code.

Introduces ll_rand() and ll_srand() with usage similar to rand(3) and srand(3)