Branch b1_4_mountconf b=4482
During an ost_add, mds_lov_sync might be calling clearorphans on all osc's
before the mds has a chance to tell the osc what it's last transno is, in
which case clearorphans resets it to 0.
So change mds_lov_sync to only sync a single target at a time; then the
clearorphans call can only sync something the mds knows about.
This also allows syncing of all osc's in parallel, and removes all
requirements for an mds_lov_desc lock in the sync.