Branch b1_4_mountconf b=4482 Problem with the mds notify method of online ost addition - lov_notify can't notify the mds until the mds is set up, due to a check of obd_set_up in obd_notify.  But obd_set_up isn't set until mds_postrecov/mds_postsetup/mds_setup returns to class_setup.
(This can be seen by adding ost's before adding the mdt - the other
order doesn't have the problem because there are no osc's in the lov
during mds_postrecov at that point, so no notifies.)

So I hacked in an additional flag check to obd_notify that happens only to
be set for the mds at this point in it's life: obd_async_recov.  Don't hurt
me.